django-app/pyproject.toml
[project]
name = "django-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1.3",
    "django-silk>=5.3.1",
    "djangorestframework>=3.15.2",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "rav>=0.0.9",
    "djangorestframework-simplejwt>=5.3.1",
]
